+++ test_integration/lib/test/test_factory.dart
@@ -1,8 +1,8 @@
 import 'package:ably_flutter_integration_test/test/appkey_provision_test.dart';
 import 'package:ably_flutter_integration_test/test/platform_and_ably_version_test.dart';
 import 'package:ably_flutter_integration_test/test/realtime_events_test.dart';
-import 'package:ably_flutter_integration_test/test/realtime_listen.dart';
 import 'package:ably_flutter_integration_test/test/realtime_publish_test.dart';
+import 'package:ably_flutter_integration_test/test/realtime_subscribe.dart';
 import 'package:ably_flutter_integration_test/test/rest_publish_test.dart';
 import 'package:ably_flutter_integration_test/test/test_helper_flutter_error_test.dart';
 import 'package:ably_flutter_integration_test/test/test_helper_unhandled_exception_test.dart';
